Bitcoin works. Shows price in PLN. Both API done.

+++ src/main/java/com/bank/dto/BitcoinDto.java
@@ -1,11 +1,14 @@
 package com.bank.dto;
 
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
 
 @Data
 @Component
@@ -14,5 +17,5 @@ import org.springframework.stereotype.Component;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class BitcoinDto {
     @JsonProperty("price")
-    double price;
+    BigDecimal price;
 }
